# apps/api/src/docs/openapi.dtos.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

export class HealthResponseDto {
  @ApiProperty({ enum: ['ok'], example: 'ok' })
  status!: 'ok';
}

export class ErrorResponseDto {
  @ApiProperty({ example: 401 })
  statusCode!: number;

  @ApiProperty({ example: 'Unauthorized' })
  message!: string | string[];

  @ApiPropertyOptional({ example: 'Unauthorized' })
  error?: string;
}

export class ApproveDraftRequestDto {
  @ApiPropertyOptional({
    description: 'Optional owner-edited message body. When omitted, the original AI draft body is sent.',
    example: '¡Hola! Sí, tenemos stock disponible. ¿Querés que te reserve uno?',
  })
  body?: string;
}

export class DraftSentResponseDto {
  @ApiProperty({ enum: ['sent'], example: 'sent' })
  status!: 'sent';
}

export class DraftRejectedResponseDto {
  @ApiProperty({ enum: ['rejected'], example: 'rejected' })
  status!: 'rejected';
}

export class CopilotQuestionRequestDto {
  @ApiProperty({
    description: 'Organization UUID whose business data should be queried.',
    example: '4e6f8a6f-9b6e-4f7f-8c1e-0f9f7c8d9a12',
  })
  organizationId!: string;

  @ApiProperty({
    description: 'Owner question for Copi.',
    example: '¿Qué clientes necesitan seguimiento hoy?',
  })
  question!: string;
}

export class OwnerCopilotResponseDto {
  @ApiProperty({
    description: 'Natural-language answer generated by Copi.',
    example: 'Tenés 7 seguimientos pendientes. Te conviene empezar por María González.',
  })
  answer!: string;

  @ApiProperty({ description: 'Server-side response time in milliseconds.', example: 42 })
  responseTimeMs!: number;

  @ApiProperty({
    description: 'Internal tools used to answer the question.',
    enum: ['messages_today', 'low_stock', 'pending_follow_ups'],
    example: ['pending_follow_ups'],
    isArray: true,
  })
  tools!: Array<'messages_today' | 'low_stock' | 'pending_follow_ups'>;
}

export class TaskMaintenanceResultDto {
  @ApiProperty({ example: 3 })
  followUpTasksCreated!: number;

  @ApiProperty({ example: 2 })
  lowStockAlertsCreated!: number;

  @ApiProperty({ example: 4 })
  pushNotificationsSent!: number;

  @ApiProperty({ example: 0 })
  pushNotificationsFailed!: number;
}

export class WhatsAppWebhookResponseDto {
  @ApiProperty({ example: true })
  received!: true;

  @ApiProperty({ example: 1 })
  eventCount!: number;
}

export class WhatsAppMessageTextDto {
  @ApiPropertyOptional({ example: 'Hola, ¿tienen stock?' })
  body?: string;
}

export class WhatsAppWebhookMessageDto {
  @ApiPropertyOptional({ example: 'wamid.HBgN...' })
  id?: string;

  @ApiPropertyOptional({ example: '5491112345678' })
  from?: string;

  @ApiPropertyOptional({ example: '1717790400' })
  timestamp?: string;

  @ApiPropertyOptional({ example: 'text' })
  type?: string;

  @ApiPropertyOptional({ type: WhatsAppMessageTextDto })
  text?: WhatsAppMessageTextDto;
}

export class WhatsAppContactProfileDto {
  @ApiPropertyOptional({ example: 'María González' })
  name?: string;
}

export class WhatsAppContactDto {
  @ApiPropertyOptional({ example: '5491112345678' })
  wa_id?: string;

  @ApiPropertyOptional({ type: WhatsAppContactProfileDto })
  profile?: WhatsAppContactProfileDto;
}

export class WhatsAppMetadataDto {
  @ApiPropertyOptional({ example: '5491123456789' })
  display_phone_number?: string;

  @ApiPropertyOptional({ example: '123456789012345' })
  phone_number_id?: string;
}

export class WhatsAppWebhookChangeValueDto {
  @ApiPropertyOptional({ example: 'whatsapp' })
  messaging_product?: string;

  @ApiPropertyOptional({ type: WhatsAppMetadataDto })
  metadata?: WhatsAppMetadataDto;

  @ApiPropertyOptional({ type: WhatsAppContactDto, isArray: true })
  contacts?: WhatsAppContactDto[];

  @ApiPropertyOptional({ type: WhatsAppWebhookMessageDto, isArray: true })
  messages?: WhatsAppWebhookMessageDto[];
}

export class WhatsAppWebhookChangeDto {
  @ApiPropertyOptional({ example: 'messages' })
  field?: string;

  @ApiPropertyOptional({ type: WhatsAppWebhookChangeValueDto })
  value?: WhatsAppWebhookChangeValueDto;
}

export class WhatsAppWebhookEntryDto {
  @ApiPropertyOptional({ example: '123456789012345' })
  id?: string;

  @ApiPropertyOptional({ type: WhatsAppWebhookChangeDto, isArray: true })
  changes?: WhatsAppWebhookChangeDto[];
}

export class WhatsAppWebhookPayloadDto {
  @ApiPropertyOptional({ example: 'whatsapp_business_account' })
  object?: string;

  @ApiPropertyOptional({ type: WhatsAppWebhookEntryDto, isArray: true })
  entry?: WhatsAppWebhookEntryDto[];
}